move redux to store
