+++ src/app/layout.tsx
@@ -4,7 +4,7 @@ import type { Metadata } from "next";
 import type { ReactNode } from "react";
 
 import { NavigationBar } from "@/components";
-import { ReduxProvider } from "@/providers/redux";
+import { ReduxProvider } from "@/providers";
 
 import { ClientLogic } from "./ClientLogic";
