+++ src/hooks/useAppDispatch.ts
@@ -1,5 +1,5 @@
 import { useDispatch } from "react-redux";
 
-import type { AppDispatch } from "@/providers/redux";
+import type { AppDispatch } from "@/store/redux";
 
 export const useAppDispatch: () => AppDispatch = useDispatch;
