+++ src/store/redux/index.ts
@@ -1,7 +1,7 @@
-export { ReduxProvider } from "./ReduxProvider";
 export type { ColorSliceState, UserSliceState } from "./slice";
 export { colorSlice, userSlice } from "./slice";
 export { setLoading, setScheme, setUser } from "./slice";
 export type { RootState } from "./store";
 export type { AppDispatch, AppThunk } from "./store";
 export { store } from "./store";
+export { ReduxProvider } from "@/providers/ReduxProvider";
