+++ src/store/redux/store.ts
@@ -1,6 +1,6 @@
 import type { ThunkAction } from "@reduxjs/toolkit";
 import { configureStore } from "@reduxjs/toolkit";
-import type { AnyAction } from "redux";
+import type { UnknownAction } from "redux";
 
 import { colorSlice, userSlice } from "./slice";
 
@@ -20,4 +20,9 @@ export const store = configureStore({
 
 export type AppDispatch = typeof store.dispatch;
 export type RootState = ReturnType<typeof store.getState>;
-export type AppThunk<ReturnType = void> = ThunkAction<ReturnType, RootState, unknown, AnyAction>;
+export type AppThunk<ReturnType = void> = ThunkAction<
+  ReturnType,
+  RootState,
+  unknown,
+  UnknownAction
+>;
